reverted to Mapbox search and display. Limited location data.
