+++ app/assets/stylesheets/application.scss
@@ -8,6 +8,8 @@
 @import "font-awesome-sprockets";
 @import "font-awesome";
 @import 'mapbox-gl/dist/mapbox-gl';
+@import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder';
+
 
 // Your CSS partials
 @import "components/index";
