+++ app/assets/stylesheets/components/_map.scss
@@ -1,3 +1,12 @@
+#map {
+  // position: absolute;
+  // // top: 0;
+  // // bottom: 0;
+  // height: 600px;
+  // width: 40%;
+}
+
+
 .mapboxgl-popup {
   max-width: 200px;
 }
