+++ app/assets/stylesheets/pages/_new_meeting.scss
@@ -8,6 +8,7 @@
 
 .new-meeting-form {
   width: 25%;
+  height: 400px;
   box-shadow: 1px 1px 5px lightgrey;
   border-radius: 5px;
   padding: 10px;
