+++ app/javascript/packs/application.js
@@ -24,55 +24,20 @@ require("channels")
 import "bootstrap";
 import { initMapbox } from '../plugins/init_mapbox';
 import { initAutocomplete } from '../plugins/init_autocomplete';
+import { initPhotonAuto } from '../plugins/init_photon_auto';
 
 document.addEventListener('turbolinks:load', () => {
   initMapbox();
 });
 
+// initPhotonAuto()
 
 // mapboxgl.accessToken = 'pk.eyJ1IjoiY2Ftcm4iLCJhIjoiY2s3dzBpYmFrMWYxcTNmcGd5NnlwdWtubyJ9.kUT1Vv1POM3rVfUfsrKSZA';
-// var map = new mapboxgl.Map({
+// const map = new mapboxgl.Map({
 //   container: 'map', // Container ID
 //   style: 'mapbox://styles/mapbox/streets-v11', // Map style to use
 //   center: [-122.25948, 37.87221], // Starting position [lng, lat]
 //   zoom: 12, // Starting zoom level
 // });
 
-// display results
-const results = document.getElementById('results');
-results.innertHTML = "";
-const insertResults = (data) => {
-  data.features.forEach((location) => {
-    // results.innerHTML = "";
-    const result = `<li>${location.properties.name}, ${location.properties.street}, ${location.properties.city}</li>`
-    results.insertAdjacentHTML('beforeend', result);
-  });
-};
-
-
-// function to query mapbox places API with the location
-const sendRequest = (query) => {
-  const photonUrl = (`http://photon.komoot.de/api/?q=${query}&limit=3`);
-  fetch(photonUrl)
-    .then(response => response.json())
-    .then((data) => {
-      console.log(data);
-      insertResults(data);
-  });
-};
-
-// grab input from location input
-const input = document.getElementById('meeting-address')
-const searchLocation = document.getElementById('search-location')
-input.addEventListener('keyup', (event) => {
-  event.preventDefault();
-  results.innerHTML = "";
-  sendRequest(input.value);
-  console.log(input.value);
-});
-
-
-// show autocomplete to fix location
-// take final location and show on map
-// save location
