+++ app/javascript/plugins/init_mapbox.js
@@ -1,4 +1,5 @@
 import mapboxgl from 'mapbox-gl';
+import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
 
 const initMapbox = () => {
   const mapElement = document.getElementById('map');
@@ -15,6 +16,8 @@ const initMapbox = () => {
       container: 'map',
       style: 'mapbox://styles/mapbox/streets-v11'
     });
+    map.addControl(new MapboxGeocoder({ accessToken: mapboxgl.accessToken,
+                                          mapboxgl: mapboxgl }));
     const markers = JSON.parse(mapElement.dataset.markers);
     markers.forEach((marker) => {
       const popup = new mapboxgl.Popup().setHTML(marker.infoWindow);
@@ -25,6 +28,10 @@ const initMapbox = () => {
     });
     fitMapToMarkers(map, markers);
   }
+
 };
 
+
 export { initMapbox };
+
+
